refactoring service controller sync() logic to be more modular
